This is the start of Note Repository project. In this version the application is able to fetch and display various notes from a MongoDB collection. Each note has it's own page.
I plan to tweak the styling of the pages some more to make them look better and implement a range of features like the ability to edit, delete and create notes.

I'm working on this simple project as a means to acquaint myself with C#, ASP.net, and the Blazor framework.

I intend to open-source this project after a little bit more work.