I have revamped the way individual content elements within each note are identified. Instead of going by the index, which would have caused issues when content elements are removed (or new ones are added) as the indexes would shift. Instead, I have created a model to represent each content element, assigning unique IDs.
This model allows me to easily track whether a content element is in edit mode or not and to keep track of the original content value, so any changes can be reverted if the edit mode is exited without saving changes.

Features added: Ability to delete and edit content elements within a note, and to delete notes themselves.

I plan to enable the creation of new notes, and the addition of new content elements. Finally, I need to fix the page flickering issue on load - likely caused by a lot of the page contents being asynchronously loaded.