Refactored code base significantly, following a domain-driven-design approach.
I have paired the models with the logic needed to manipulate them. This is more in-line with what a domain model should be.

I have created a notes service to handle interacting with the database layer. Within this service, the note-related models passed into it are translated into simple DTOs.

Noticing that the UI would not automatically update when deleting a note from the list of notes. I created events that are listened for by the UI (razor) components which will trigger a refresh.

These events are emitted by the notes service, indicating a change at the database level.

Similarly, I explored various ways to dynamically include a particular component within the MainLayout file. The solution I ultimately settled on was using a dynamic component which would acquire what button it should be from a service.

This service is populated with the appropriate button type by the page currently rendered. E.g., on the Note page the button should enable the creation of a new 'Content' field.

An event is fired by the button service which is listened to by the layout, this causes the button to actually render properly.

While an arguably complicated approach, I think it looks tidier than having a number of If statements to render buttons conditionally depending on the URL detected.

Functionally the application can now be used to view, delete and add notes, and also edit the contents within each.

I still need to add the 'CreateContentButton', after which I will mainly work towards styling the site.